+++ js/src/main/scala/spatutorial/client/components/package.scala
@@ -1,0 +1,7 @@
+package spatutorial.client
+
+import scala.scalajs.js
+
+package object components extends js.GlobalScope {
+  val jQuery: JQueryStatic = js.native
+}