+++ project/build.scala
@@ -20,7 +20,7 @@ object SPABuild extends Build {
   lazy val spa = crossProject.in(file(".")).
     settings(
       name := "scalajs-spa",
-      version := "0.1.1",
+      version := "0.1.2",
       scalaVersion := "2.11.5",
       scalacOptions ++= Seq(
         "-Xlint",