+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Change history
 
+## 0.1.2
+
+* Simple jQuery integration added (Bootstrap Modal)
+* Modal example with a form
+* Refactored Bootstrap components a bit
+* Todos are now updated on the server
+
 ## 0.1.1
 
 * Refactored the router system to follow the intended design of the Scala.js React router (thanks to @japgolly for feedback)